requirements.txt
aiohttp==3.6.2
apiai==1.2.3
async-timeout==3.0.1
attrs==19.2.0
chardet==3.0.4
idna==2.8
multidict==4.5.2
numpy==1.17.2
slackclient==2.2.1
yarl==1.3.0
